Fix YamlReader ignore sequences when ignoreUnknownProperties set
When there was no List in target class and YAML input had a sequence,
only the SCALAR event type was ignored resulting in passing
SEQUENCE_START for the next iteration and cast error.

Now when ignoring a property the code is checking if the next event
is SEQUENCE_START. If yes, it goes through it to the corresponding
SEQUENCE_END event.